refactor(metadata): one lock store for both SQL dialects
sqlite and postgres each carried a full lock store and a second, near-
identical copy of every method for the transaction path — 1,443 lines
where the only per-call difference was which executor ran the statement.

Core already is an executor plus a dialect, and both the store and its
transaction embed one, so the lock bodies move onto Core and both paths
fall out of that: the 24 transaction variants and the 24 delegates are
gone rather than merged.

What stays per-dialect is what genuinely differs — seven statements that
disagree on placeholder syntax or CURRENT_TIMESTAMP versus NOW(), and
the WHERE tail ListLocks assembles at run time, which is the one lock
query with no constant to spell. Three statements carry neither and are
shared outright.

Also states lock.LockStore as a compile-time assertion on each store:
metadata.Store does not embed it, and the only caller reaches it through
a type assertion that skips lock setup silently on a miss.

Verified against a real postgres as well as sqlite, and both dialects'
conformance runs were checked against deliberately broken builds — a
swapped delete-by-client/by-file pair and an off-by-one in the postgres
placeholder numbering — so the suite is known to see this code.
