+++ pkg/metadata/store/sqlite/store.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/marmos91/dittofs/internal/logger"
 	"github.com/marmos91/dittofs/pkg/metadata"
+	"github.com/marmos91/dittofs/pkg/metadata/lock"
 	"github.com/marmos91/dittofs/pkg/metadata/store/basestore"
 	"github.com/marmos91/dittofs/pkg/metadata/store/internal/sharecache"
 
@@ -91,9 +92,8 @@ type SQLiteMetadataStore struct {
 	// thereafter. Immutable for the life of the instance.
 	storeID string
 
-	// Sub-stores for lock / client / durable-handle / NFSv4-recovery
-	// persistence. Each wraps the shared *sql.DB executor.
-	lockStore     *sqliteLockStore
+	// Sub-stores for client / durable-handle / NFSv4-recovery persistence.
+	// Each wraps the shared *sql.DB executor.
 	clientStore   *storesql.ClientStore
 	durableStore  *sqliteDurableStore
 	recoveryStore *storesql.RecoveryStore
@@ -168,7 +168,6 @@ func NewSQLiteMetadataStore(
 
 	// The substores derive only from db, which is never reassigned, so bind
 	// them once here.
-	store.lockStore = newSQLiteLockStore(store.conn())
 	store.clientStore = &storesql.ClientStore{X: store.conn(), D: sqliteDialect}
 	store.durableStore = newSQLiteDurableStore(store.conn())
 	store.recoveryStore = &storesql.RecoveryStore{X: store.conn(), D: sqliteDialect}
@@ -389,3 +388,9 @@ func (s *SQLiteMetadataStore) RecomputeUsage(ctx context.Context) error {
 	s.quotaMu.Unlock()
 	return s.initUsedBytesCounter(ctx)
 }
+
+// metadata.Store does not embed lock.LockStore: the store's lock surface is
+// reached through a runtime type assertion, which skips lock initialisation
+// silently rather than failing when the store no longer satisfies it. This
+// states the requirement where the compiler can see it.
+var _ lock.LockStore = (*SQLiteMetadataStore)(nil)
